Fix compilation warnings; OCaml 5 volatile correctness
This patch fixes a number of compilation warnings that arose when
compiling `pcre2_stubs.c` under Apple Clang 14.0.3 with the standard
compilation flags as defined in the `dev` dune `env`ironment.
Additionally, in so doing, it resolves a number of `volatile` cast
warnings that could cause memory consistency issues when paired with the
OCaml 5 runtime.